chore: separate optimization and transformation
